Build .html and institute getComputerChoice function
Wrote boilerplate in index.html and linked script.js file.  Then built function getComputerChoice to establish the Computer's play for the rock paper scissors game.  Right now, it returns a string, it may be better if it returns an integer that we can use to compare the user's play to the computer play.